api/throttling.py
from rest_framework.throttling import AnonRateThrottle, UserRateThrottle


class BurstRateThrottle(UserRateThrottle):
    scope = 'burst'


class SustainedRateThrottle(UserRateThrottle):
    scope = 'sustained'


class AnonBurstRateThrottle(AnonRateThrottle):
    scope = 'anon_burst'


class AnonSustainedRateThrottle(AnonRateThrottle):
    scope = 'anon_sustained'
